Move the `grid` module into `src`
So it can be used outside storybook.
